refactor notification (working)
