+++ JokeAbout/data/src/main/java/com/soldatov/jokeabout/data/service/workmanager/JokeWorker.kt
@@ -8,6 +8,12 @@ import com.soldatov.jokeabout.data.api.models.JokeRequestData
 import com.soldatov.jokeabout.data.repository.JokeRepositoryImpl
 import com.soldatov.jokeabout.data.service.alarmmanager.JokeAlarmService
 import com.soldatov.jokeabout.data.storage.SharedPrefJokeStorage
+import com.soldatov.jokeabout.data.utils.JokeNotification
+
+const val CHANNEL_ID = "uniqueId"
+const val NOTIFICATION_ID = 101
+const val NOTIFICATION_TITLE = "New joke"
+const val ACTIVITY_TO_START = "com.soldatov.jokeabout.presentation.JokeAboutActivity"
 
 class JokeWorker(context: Context, workerParameters: WorkerParameters) :
     Worker(context, workerParameters) {
@@ -20,7 +26,10 @@ class JokeWorker(context: Context, workerParameters: WorkerParameters) :
         val jokeRequestData = JokeRequestData(data.firstName, data.lastName)
         val jokeApiImpl = JokeApiImpl()
         val joke = jokeApiImpl.getJoke(applicationContext, jokeRequestData)
-        // FIXME здесь должна отправляться нотификация
+        val jokeNotification = JokeNotification()
+        if (joke != null) {
+            jokeNotification.showNotification(applicationContext, joke.value.joke)
+        }
         return Result.success()
     }
